Refactor Project Metadata into own component
